creating api for adding item into inventory and delete receive item

+++ app/services/inventory_item.py
@@ -43,19 +43,19 @@ class InventoryItemService(CommonService):
         self.db = db
         CommonService.__init__(self,db, InventoryItemModel)
 
-    def add_item_in_inventory(self,po_id, py_model:BaseModel):
+    def add_item_in_inventory(self,py_model:BaseModel):
         pydantic_data = py_model.model_dump()
         logger.info(f"Adding item in inventory with data: {pydantic_data}")
 
-        purchase_order_record = self.db.get(PurchaseOrderModel, po_id)
+        purchase_order_record = self.db.get(PurchaseOrderModel, pydantic_data['po_id'])
 
         #raise exception if purchase_order_record not found
         if not purchase_order_record:
-            logger.warning(f"Purchase Order with id {po_id} not exists")
-            raise NotFoundException(f"Purchase Order with id {po_id} not exists")
+            logger.warning(f"Purchase Order with id {pydantic_data['po_id']} not exists")
+            raise NotFoundException(f"Purchase Order with id {pydantic_data['po_id']} not exists")
         
         if purchase_order_record.status != PurchaseOrderStatus.RECEIVED:
-            logger.warning(f"Purchase Order with id {po_id} status is not received")
+            logger.warning(f"Purchase Order with id {pydantic_data['po_id']} status is not received")
             raise InvalidStatusException(f"Cannot proceed. Purchase Order must be in 'RECEIVED' state, currently in {purchase_order_record.status.value}")
         
 
@@ -65,6 +65,10 @@ class InventoryItemService(CommonService):
         if not received_item_record:
             logger.warning(f"Received item with product id{pydantic_data['product_id']} not found!")
             raise NotFoundException(f"Received item with product id {pydantic_data['product_id']} not found!")
+        
+        if pydantic_data['qty'] > received_item_record.qty:
+            logger.warning(f"Product with id {pydantic_data['product_id']} does not receive {pydantic_data['qty']}. Actual unit is : {received_item_record.qty}")
+            raise NotFoundException(f"Product with id {pydantic_data['product_id']} does not receive {pydantic_data['qty']} units. Actual unit is : {received_item_record.qty}")
         
         bin_record = self.db.query(BinModel).filter_by(id = pydantic_data['bin_id'], warehouse_id = purchase_order_record.warehouse_id).first()
 
@@ -78,16 +82,26 @@ class InventoryItemService(CommonService):
             logger.warning(f"The quantity ({pydantic_data['qty']}) exceeds the bin's available capacity ({bin_record.max_units})")
             raise BinCapacityExceededException(bin_record.max_units, pydantic_data['qty'])
         
+        #add qty to bin current stock
         bin_record.current_stock_units = bin_record.current_stock_units + pydantic_data['qty']
+        #substract qty from available unit
         bin_record.available_units = bin_record.available_units - pydantic_data['qty']
 
         pydantic_data["product_id"] = received_item_record.product_id
+        pydantic_data["sku"] = received_item_record.sku
+        pydantic_data["warehouse_id"] = purchase_order_record.warehouse_id
         logger.info(f"Adding item in inventory with data: {pydantic_data}")
+        pydantic_data.pop("po_id")
+        logger.info(f"Deleting po_id from pydantic data: {pydantic_data}")
         inventory_item = self.create_record(pydantic_data)
+        logger.info(f"Deleting received item with id {received_item_record.id}")
 
+        #substract received item qty which is added into bin
+        received_item_record.qty = received_item_record.qty - pydantic_data['qty']
 
         try:
-            self.db.delete(received_item_record)
+            if received_item_record.qty == 0:
+                self.db.delete(received_item_record)
             self.db.commit()
         except SQLAlchemyError as e:
             self.db.rollback()
